Crate agents

AlGlobo Agents - Process payments

This program sets up the various agents in the agents.yaml config file so that they can be used to process flight payments.

Start the program with cargo run --bin agents

Each agent will be listening on the configured TCP port, and will log and return the transaction states.

The agents.yaml file is defined as a list of items like

- name: "bank" // the name of the agent
  successrate: 0.9 // the rate on which they accept payments
  port: 1024 // the port to listen

Modules

agent

Agent Struct

communication

Several definitions regarding the alglobo<->agents communication protocol

logger

Logging functions

utils

Global utils related to reading and parsing of csv and yaml files

Functions

create_listener

Constantly listens to TCP connections Handles different 2-phase transaction messages like PREPARE and COMMIT Stops listening on a F

main

Main function. Starts the agents from the .yaml configuration file and the agent killer. Finishes when all the agents are killed.

psycho_agent_killer

Starts the agent killer in a new thread, killing agents via keyboard input